Added more tests Added double headed axe Added hoe Added shovel Improved assembly table

+++ fabric/src/main/java/com/sigmundgranaas/forgero/client/model/Unbaked2DTexturedModel.java
@@ -83,8 +83,8 @@ public class Unbaked2DTexturedModel implements UnbakedFabricModel {
         for (ModelElementFace face : element.faces.values()) {
             var offset = offsetMap.getOrDefault(face.textureId, new Offset(0, 0));
             if (offset.x() != 0 || offset.y() != 0) {
-                element.from.add((float) offset.x(), (float) offset.y(), 0.001f);
-                element.to.add((float) offset.x(), (float) offset.y(), 0.001f);
+                element.from.add(offset.x(), offset.y(), 0.001f);
+                element.to.add(offset.x(), offset.y(), 0.001f);
             }
             break;
         }
